order-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
async-timeout==4.0.2
    # via redis
click==8.1.3
    # via
    #   flask
    #   rq
croniter==1.3.8
    # via rq-scheduler
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-rq2
flask-rq2==18.3
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
pyjwt==2.6.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   croniter
    #   rq-scheduler
redis==4.4.1
    # via
    #   -r requirements.in
    #   flask-rq2
    #   rq
rq==1.11.1
    # via
    #   -r requirements.in
    #   flask-rq2
    #   rq-scheduler
rq-scheduler==0.11.0
    # via flask-rq2
six==1.16.0
    # via python-dateutil
werkzeug==2.2.2
    # via flask
